requirements.txt
# Core Web Framework
Flask>=2.0.1
Flask-Cors>=3.0.10
gunicorn>=20.1.0
Werkzeug>=2.0.0

# Data Processing & Analysis
numpy>=1.22.0
pandas>=2.0.0
scipy>=1.7.0
scikit-learn>=1.0.2

# Image Processing
Pillow>=10.0.0
opencv-python>=4.7.0
torchvision>=0.12.0
torch>=1.11.0

# AI & Machine Learning
google-generativeai>=0.3.0
tensorflow>=2.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# System Monitoring & Performance
psutil>=5.8.0
py-cpuinfo>=8.0.0

# Logging & Monitoring
python-json-logger>=2.0.0
structlog>=21.1.0

# Security
python-dotenv>=0.19.0
cryptography>=3.4.7
PyJWT>=2.3.0

# Testing
pytest>=7.0.0
pytest-cov>=2.12.0
pytest-mock>=3.6.1

# Development Tools
black>=22.0.0
flake8>=4.0.0
mypy>=0.910
isort>=5.9.3

# Documentation
Sphinx>=4.2.0
sphinx-rtd-theme>=1.0.0

# Database
SQLAlchemy>=1.4.0
alembic>=1.7.0

# Caching
Flask-Caching>=1.10.0
redis>=4.0.0

# API
requests>=2.31.0
aiohttp>=3.8.0
httpx>=0.23.0

# Performance Monitoring
prometheus-client>=0.12.0
statsd>=3.3.0

# Error Tracking
sentry-sdk>=1.5.0

# Async Support
aiofiles>=0.8.0
asyncio>=3.4.3

# Date/Time Handling
pytz>=2021.3
python-dateutil>=2.8.2

# File Handling
python-magic>=0.4.24
filetype>=1.0.0

# Compression
python-snappy>=0.6.0
